+++ ui/src/Header/SpaceSelectionTabs.scss
